requirements.txt
Django==5.0.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
openai==1.38.0
python-decouple==3.8